+++ src/background/error_handler.rs
@@ -15,12 +15,15 @@ macro_rules! define_app_errors {
 
 #[macro_export]
 macro_rules! log_error {
-    ($($result:expr),*) => {
-        $(
-            if let Err(err) = $result {
-                log::error!("{:?}", err);
-            }
-        )*
+    ($result:expr) => {
+        if let Err(err) = $result {
+            log::error!("{:?}", err);
+        }
+    };
+    ($result:expr, $context:expr) => {
+        if let Err(err) = $result {
+            log::error!("Context: {:?} Err: {:?}", $context, err);
+        }
     };
 }