+++ src/background/seelen_bar/hook.rs
@@ -1,5 +1,3 @@
-use windows::Win32::Foundation::HWND;
-
 use crate::{
     error_handler::Result,
     windows_api::{window::Window, WindowsApi},
@@ -9,24 +7,23 @@ use crate::{
 use super::FancyToolbar;
 
 impl FancyToolbar {
-    pub fn process_win_event(&mut self, event: WinEvent, origin: HWND) -> Result<()> {
-        let window = Window::from(origin);
+    pub fn process_win_event(&mut self, event: WinEvent, window: &Window) -> Result<()> {
         match event {
             WinEvent::ObjectNameChange => {
-                if self.last_focus == Some(origin) {
-                    self.focus_changed(origin)?;
+                if self.last_focus == Some(window.hwnd()) {
+                    self.focus_changed(window.hwnd())?;
                 }
             }
             WinEvent::SystemForeground | WinEvent::ObjectFocus => {
-                self.focus_changed(origin)?;
-                self.handle_overlaped_status(origin)?;
+                self.focus_changed(window.hwnd())?;
+                self.handle_overlaped_status(window.hwnd())?;
             }
             WinEvent::ObjectLocationChange => {
                 if window.hwnd() == self.hwnd()? {
                     self.set_position(window.monitor().handle())?;
                 }
-                if origin == WindowsApi::get_foreground_window() {
-                    self.handle_overlaped_status(origin)?;
+                if window.is_focused() {
+                    self.handle_overlaped_status(window.hwnd())?;
                 }
             }
             WinEvent::SyntheticFullscreenStart(event_data) => {