+++ changelog.md
@@ -20,6 +20,7 @@
 ### fix
 - showing domain on username for local accounts.
 - showing application frame host instead real app name.
+- incorrect event order on win events.
 
 ## [2.1.9]
 ### fix